set virtual-physical mapping in TranspileResult
